Define wrapper model classes to avoid duplicate code in models
